Update users.mdx
Proposing to remove the bottom section as engineering per this thread says that this option isn't available in TFC.

https://hashicorp.slack.com/archives/C7HG09LBV/p1653073595938819

remove
Note: When SAML SSO is enabled there is a session timeout for user API tokens, forcing users to periodically reauthenticate through the web UI in order to keep their tokens active. Refer to API Token Expiration for details.