Add a table to show the full list of supported tfce resources
We also support a long list of resources that do not have cost associated with them. They don't show up in the totals but we correctly ignore them in our totals. This is different to other resources because we don't know if those would have cost or not, so they're "unknown" rather than "ignored".