Switch blue diamond for circle in token type table
Seeing the page as is felt a bit odd to me. It's pretty clear, but I wondered if it was colorblind-accessible.
I ran the page through a few colorblind filters, and it seemed okay. An achromatopsia filter (greyscale) is kinda unclear, but might be okay.
But I'm also aware that I'm seeing the macOS versions of these emojis, and other emoji fonts might not work so well.

With different shapes, the difference between the icons is much clearer)

I am not colorblind myself, so this may be entirely unnecessary, but as we don't use these symbols anywhere else in the docs, it seemed like a safe enough change to do regardless.