Runs list API: filter params can accept comma-separated lists of values
These aren't restricted to single values; you can throw in a bunch of acceptable
values to get an "OR" search. I validated this by searching for
`filter[status]=discarded,cancelled` on a workspace with 12 discarded runs and 1
cancelled one, and got the expected 13 runs in my results. To validate the
intended behavior in the atlas code, check the `Api::V2::RunsIndexFilter`
service object's filter methods and note that they split their input on `,`.