Update entitlement set, feature set, subscription and policy API endpoints after new free tier (#355)
* Update entitlement sets API for task and policy limits

The new free tier within TFC has added some new properties to the entitlement set
API. This commit updates the documentation with the new properties in the response.

* Update feature sets API for task and policy limits

The new free tier within TFC has added some new properties to the feature set
API. This commit updates the documentation with the new properties in the response.

* Update subscription API for task and policy limits

The new free tier within TFC has added some new properties to the subscription
API. This commit updates the documentation with the new properties in the response.

* Update the Policies API with new and deprecated properties

The Policies API has been changed to deprecate the old `enforce` property in favour
of the new `enforcement-level` property. This commit updates the Policies API documentation
to instruct API consumers to use the new property.

* Update API changelog for free tier limits

This commit updates the API changelog with all of the API updates for the free tier
limits.