Added a reference to provider configuration in Test Integrated Modules
I was doing research into Test Integrated Modules for a recent project, and it took me a little while to work out how to provide a test-specific provider configuration as there is no mention of this in the registry docs. Have added a link to the specific section of the terraform test docs that deal with this.