cloud-docs/run-environment: Don't overload the term "run" (#638)
* cloud-docs/run-environment: Don't overload the term "run"

The "run environment" page belongs to a collection of pages that describe
the capital-R concept of "Run" in HCP Terraform, which refers to the
combination of a plan phase and an optional apply phase. "Normal" runs
form a per-workspace queue where each run must be resolved (its plan either
approved or declined) before proceeding with the next.

This particular section, however, was using "run" both in its usual meaning
_and_ with a new meaning of "job that can be picked up by a worker", which
made this section hard to follow when considered in the broader context
where it's placed.

The prior text was already using the word "job" to describe what was being
taken from queues, and so here I've revised the text to consistently use
that noun to refer to the kind of thing that is constrained by our
concurrency limits. I think that explains the situation a little more
clearly while avoiding getting _too_ deep into the implementation details.

I also made the separate editorial decision that jobs take actions and so
what they do should be described as active agents, e.g. using the gerund
"applying" instead of the nouned-verb "applies" (which is ambiguous with
the present-tense verb as in "the agent _applies_ the plan".

Otherwise I attempted to keep the message largely the same as it was
before, but just using hopefully-clearer terminology.

* Apply suggestions from code review

* Apply suggestions from code review

---------

Co-authored-by: trujillo-adam <47586768+trujillo-adam@users.noreply.github.com>